Prototype for Import
Starting point for importing classes, Creating a new class manually
does not work as of this commit however, you can import a class using
the “import.csv” file then click “Import Class List” and it should work
fine.  Might need to also edit the links to point to the jquery listed
in the repo but I will do that next commit.