Clock in complete
Finished clocking in. Reworked it to instead of using a separate table
to create clock in sessions with multiple courses it instead creates a
unique time clock record for each class you clock in to. Made it so when
you are clocked into a class if you go back to the clock in page that
class option is disabled on the form selection.