Clock out
Finished clock out. Only lets you clock out of classes you are clocked
in. Updated the navbar to include clock out, only when clocked in.
Updated checklogin to also set the Session variable for active to check
if you have any clocked in classes when you log in if a session was
destroyed.